Move ParticleFlow Alpaka data formats to separate sub-package; this should undone eventually
