--- DataFormats/ParticleFlowReco/src/alpaka/classes_cuda.h
@@ -1,7 +1,0 @@
-#include "DataFormats/Common/interface/DeviceProduct.h"
-#include "DataFormats/Common/interface/Wrapper.h"
-#include "DataFormats/Portable/interface/Product.h"
-#include "DataFormats/ParticleFlowReco/interface/PFRecHitSoA.h"
-#include "DataFormats/ParticleFlowReco/interface/alpaka/PFRecHitDeviceCollection.h"
-#include "DataFormats/ParticleFlowReco/interface/CaloRecHitSoA.h"
-#include "DataFormats/ParticleFlowReco/interface/alpaka/CaloRecHitDeviceCollection.h"
