--- DataFormats/ParticleFlowReco/src/alpaka/classes_serial.h
@@ -1,6 +1,0 @@
-#include "DataFormats/Common/interface/Wrapper.h"
-#include "DataFormats/Portable/interface/Product.h"
-#include "DataFormats/ParticleFlowReco/interface/PFRecHitSoA.h"
-#include "DataFormats/ParticleFlowReco/interface/PFRecHitHostCollection.h"
-#include "DataFormats/ParticleFlowReco/interface/CaloRecHitSoA.h"
-#include "DataFormats/ParticleFlowReco/interface/CaloRecHitHostCollection.h"
