+++ DataFormats/ParticleFlowReco/src/classes_1.h
@@ -1,0 +1,54 @@
+#include "DataFormats/Common/interface/Wrapper.h"
+#include "DataFormats/Common/interface/Ref.h"
+#include <DataFormats/Common/interface/OwnVector.h>
+#include <DataFormats/Common/interface/ClonePolicy.h>
+#include "DataFormats/Common/interface/ValueMap.h"
+
+#include "DataFormats/ParticleFlowReco/interface/PFCluster.h"
+#include "Math/Cartesian3D.h"
+#include "DataFormats/ParticleFlowReco/interface/PFRecHit.h"
+#include "Math/Polar3D.h"
+#include "Math/CylindricalEta3D.h"
+#include "DataFormats/ParticleFlowReco/interface/PFRecTrack.h"
+#include "DataFormats/ParticleFlowReco/interface/GsfPFRecTrack.h"
+#include "Math/GenVector/PositionVector3D.h"
+#include "DataFormats/Math/interface/Point3D.h"
+#include "Rtypes.h"
+#include "DataFormats/Math/interface/Vector3D.h"
+#include "Math/PxPyPzE4D.h"
+#include "DataFormats/DetId/interface/DetId.h"
+#include "DataFormats/TrackReco/interface/Track.h"
+#include "DataFormats/TrackReco/interface/TrackFwd.h"
+#include "DataFormats/GsfTrackReco/interface/GsfTrack.h"
+#include "DataFormats/GsfTrackReco/interface/GsfTrackExtra.h"
+#include "DataFormats/ParticleFlowReco/interface/PFSimParticle.h"
+#include "DataFormats/ParticleFlowReco/interface/PFTrajectoryPoint.h"
+#include "DataFormats/ParticleFlowReco/interface/PFBlockElement.h"
+#include "DataFormats/ParticleFlowReco/interface/PFBlockElementTrack.h"
+#include "DataFormats/ParticleFlowReco/interface/PFBlockElementGsfTrack.h"      //Daniele
+#include "DataFormats/ParticleFlowReco/interface/PFBlockElementBrem.h"          //Daniele
+#include "DataFormats/ParticleFlowReco/interface/PFBlockElementSuperCluster.h"  //Florian
+#include "DataFormats/ParticleFlowReco/interface/PFBlockElementCluster.h"
+#include "DataFormats/ParticleFlowReco/interface/PFBlock.h"
+#include "DataFormats/ParticleFlowReco/interface/PFDisplacedTrackerVertex.h"
+#include "DataFormats/ParticleFlowReco/interface/PFNuclearInteraction.h"
+#include "DataFormats/ParticleFlowReco/interface/PFConversion.h"
+#include "DataFormats/ParticleFlowReco/interface/PFConversionFwd.h"
+#include "DataFormats/ParticleFlowReco/interface/PFV0.h"
+#include "DataFormats/ParticleFlowReco/interface/ConvBremSeed.h"
+#include "DataFormats/ParticleFlowReco/interface/ConvBremSeedFwd.h"
+#include "DataFormats/ParticleFlowReco/interface/ParticleFiltrationDecision.h"
+
+#include "DataFormats/ParticleFlowReco/interface/PFDisplacedVertexCandidate.h"
+#include "DataFormats/ParticleFlowReco/interface/PFDisplacedVertexSeed.h"
+#include "DataFormats/ParticleFlowReco/interface/PFDisplacedVertex.h"
+#include "DataFormats/ParticleFlowReco/interface/PFDisplacedVertexCandidateFwd.h"
+#include "DataFormats/ParticleFlowReco/interface/PFDisplacedVertexSeedFwd.h"
+#include "DataFormats/ParticleFlowReco/interface/PFDisplacedVertexFwd.h"
+
+#include "DataFormats/ParticleFlowReco/interface/PreId.h"
+#include "DataFormats/ParticleFlowReco/interface/PreIdFwd.h"
+#include "DataFormats/ParticleFlowReco/interface/RecoPFClusterRefCandidate.h"
+#include "DataFormats/ParticleFlowReco/interface/RecoPFClusterRefCandidateFwd.h"
+
+#include <map>
