+++ DataFormats/ParticleFlowReco_Alpaka/interface/CaloRecHitHostCollection.h
@@ -2,7 +2,7 @@
 #define ParticleFlowReco_CaloRecHitHostCollection_h
 
 #include "DataFormats/Portable/interface/PortableHostCollection.h"
-#include "DataFormats/ParticleFlowReco/interface/CaloRecHitSoA.h"
+#include "DataFormats/ParticleFlowReco_Alpaka/interface/CaloRecHitSoA.h"
 
 using CaloRecHitHostCollection = PortableHostCollection<CaloRecHitSoA>;
