+++ DataFormats/ParticleFlowReco_Alpaka/interface/PFRecHitHostCollection.h
@@ -1,7 +1,7 @@
 #ifndef ParticleFlowReco_PFRecHitHostCollection_h
 #define ParticleFlowReco_PFRecHitHostCollection_h
 
-#include "DataFormats/ParticleFlowReco/interface/PFRecHitSoA.h"
+#include "DataFormats/ParticleFlowReco_Alpaka/interface/PFRecHitSoA.h"
 #include "DataFormats/Portable/interface/PortableHostCollection.h"
 
 using PFRecHitHostCollection = PortableHostCollection<PFRecHitSoA>;
