+++ DataFormats/ParticleFlowReco_Alpaka/interface/alpaka/CaloRecHitDeviceCollection.h
@@ -1,7 +1,7 @@
 #ifndef ParticleFlowReco_CaloRecHitDeviceCollection_h
 #define ParticleFlowReco_CaloRecHitDeviceCollection_h
 
-#include "DataFormats/ParticleFlowReco/interface/CaloRecHitSoA.h"
+#include "DataFormats/ParticleFlowReco_Alpaka/interface/CaloRecHitSoA.h"
 #include "DataFormats/Portable/interface/alpaka/PortableCollection.h"
 #include "HeterogeneousCore/AlpakaInterface/interface/config.h"
