+++ DataFormats/ParticleFlowReco_Alpaka/src/alpaka/classes_serial.h
@@ -1,0 +1,6 @@
+#include "DataFormats/Common/interface/Wrapper.h"
+#include "DataFormats/Portable/interface/Product.h"
+#include "DataFormats/ParticleFlowReco_Alpaka/interface/PFRecHitSoA.h"
+#include "DataFormats/ParticleFlowReco_Alpaka/interface/PFRecHitHostCollection.h"
+#include "DataFormats/ParticleFlowReco_Alpaka/interface/CaloRecHitSoA.h"
+#include "DataFormats/ParticleFlowReco_Alpaka/interface/CaloRecHitHostCollection.h"
