+++ DataFormats/ParticleFlowReco_Alpaka/src/classes.h
@@ -1,0 +1,2 @@
+#include "DataFormats/ParticleFlowReco_Alpaka/interface/PFRecHitSoA.h"
+#include "DataFormats/ParticleFlowReco_Alpaka/interface/CaloRecHitSoA.h"
