+++ RecoParticleFlow/PFRecHitProducer/plugins/BuildFile.xml
@@ -6,6 +6,7 @@
   <use name="HeterogeneousCore/AlpakaTest"/>
   <use name="DataFormats/Common"/>
   <use name="DataFormats/ParticleFlowReco"/>
+  <use name="DataFormats/ParticleFlowReco_Alpaka"/>
   <use name="CommonTools/ParticleFlow"/>
   <use name="DQMServices/Core"/>
   <flags EDM_PLUGIN="1"/>
@@ -15,6 +16,7 @@
 <library name="RecoParticleFlowPFRecHitProducersPluginsPortable" file="alpaka/*.cc">
   <use name="alpaka"/>
   <use name="DataFormats/ParticleFlowReco"/>
+  <use name="DataFormats/ParticleFlowReco_Alpaka"/>
   <use name="FWCore/Framework"/>
   <use name="FWCore/MessageLogger"/>
   <use name="FWCore/ParameterSet"/>
